+++ CHANGELOG.md
@@ -15,6 +15,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 
 ## [0.1.x]
 
+- Added alternative max-friendly style for `add` such that `<filename> arg1 arg2 .. argN` is interpretered as `<filename>:arg1:arg2:argN`
+
+- Added tested for chuck `add` colon-separated args
+
 - Added `ck_info`, `ck_warn`, `ck_debug` for more control and consistency wrt console logging.
 
 - Removed `info` message and merged its functionality with `status` message.